Datentyp für QR-Code-Nummer von int auf long geändert

+++ API.java
@@ -1,3 +1,3 @@
 public interface API {
-    public Lieferschein getLieferschein(int qrNumber);
+    public Lieferschein getLieferschein(long qrNumber);
 }
